Created an ipp-usb Snap
As part of an all-Snap printing experience we have also a Snap of
ipp-usb now, especially to support IPP-over-USB in an all-Snap OS
distribution, to add IPP-over-USB support if the distribution does not
support it, or to get the newest ipp-usb onto an already released
distribution.

Instructions and technical background are in the README.md file.