Add quirk to deny certain device capabilities
Most devices annouce a whole set of capabilities print,scan,fax
and users wants to choose to disable some of them because their
either broken or some other reason.

It fixes an issue where ipp-usb would reset a device usb connection
every few seconds because one of the caps is broken on the device.