Add quirk to disable fax
Most devices annouce a whole set of capabilities print,scan,fax
and users might choose to disable fax because it is
either broken or they don't have any use for it.

It also fixes an issue where ipp-usb would
reset a device usb connection
every few seconds because fax is broken on the device.