Snap: Added support for "ipp-usb status"/"ipp-usb check" calls
Before, the Snap had only one app, the UDEV-rule replacement daemon
script "run-ipp-usb", running the ipp-usb executable on its own was
not possible in the Snap.

As now ipp-usb has two non-daemon run modes, "ipp-usb status" to check
ststus of a running daemon and "ipp-usb check" to check presence of
potentially supported USB devices, we need now a second app, to run
ipp-usb in these modes.

For this we rename the current "run-ipp-usb" script into
"run-ipp-usb-server" and also the current "ipp-usb" app in
snapcraft.yaml into "ipp-usb-server", to free the "ipp-usb" name for
the non-daemon mode.

Then we create a new "run-ipp-usb" script for the non-daemon mode of
ipp-usb, being run by calling the newly created "ipp-usb" app. The
script does the same directory and file copy preparation as the server
script but it does not overwrite any files and ignores errors (usually
"Permission denied" when running as non-root). It then simply calls
the ipp-usb executable with the command line arguments the user
supplied in the original call. The "ipp-usb" app plugs "usb-raw" and
"hardware-observe" to access the USB for the "check" mode.

This way we can call

    ipp-usb status

and

    ipp-usb check

on systems with the ipp-usb Snap installed now and the two commands
will do the right thing now, the former will always access the snapped
daemon, the latter can need root privileges to access the USB,
depending on the operating system's configuration.