Pantum M7300FDW support
Pantum M7300FDW encodes IPP messages improperly. It violates collection
encoding rules. As a workaround, support of broken IPP messages was
added to the github.com/OpenPrinting/goipp library.

Here we upgrade to the latest version of github.com/OpenPrinting/goipp
(not yet released) and enable decoder workarounds.

Please note, to get IPP printing working with this device, the similar
change needs to be implemented in CUPS, which is not done yet.

Scanning not affected, it was fully functional before.