Use flock() instead of lockf() for file locking
lockf() is per-process and it doesn't allow one thread of process to
wait while another thread of the same process holds the file lock (it
returns EDEADLK instead of blocking).

It makes this kind of locking useless for multithreaded program. So
lockf() was replaced with flock().

This change is required for #77