auth: don't attempt to resolve UID, if UID is not known
Client UID may be unknown in many situations (for example,
non-local network connection from a client).

Attempt to resolve such an unknown UID will definitely fail
and should not be resolved.

Documentation promises that user name will be exposed as "*"
for rule matching. This change brings the real behavior to the
documented one.