auth: don't try to obtain client UID, if we don't actually use it
For example, if no auth rules depend on UID (i.e., authentication
is not configured, or all rules are wildcard), we don't need to obtain
an actual UID.

In some containerized installation (for example, in current snap
package), sock_diag access is not permitted, which leads ipp-usb
to fail even if actually authentication is not requested.

This change works around this issue.