Removed debian/ subdirectory
The debian/ directory makes maintaining Debian (and Ubuntu) packages
overly complicated. The debian/ directory is for the Debian packaging
and is maintained by the Debian and Ubuntu package maintainers. Having
a debian/ directory in the upstream source, even if it contains the
files to create a valid Debian package, causes conflicts with the
debian/ directory of Debian and Ubuntu. Therefore we are removing it
here.