Added 'zlp-send'
This quirk, if enabled, sets LIBUSB_TRANSFER_ADD_ZERO_PACKET flag on
each outgoing transfer, which causes libusb to terminate outgoing
transfers that a multiple of the endpoint's packet size win an
extra zero length packet.

Note, there is actually no known device that actually requires this
quirk to be set. It is added just for for completeness.