Code formatting updated
Golang tools (editor plugins etc) enforce world-wide uniform code formatting
rules. Unfortunately, these rules sometimes change, though rarely and not
significantly.

The drawback of that is when somebody edits a file, initially created with
the old tools version, the planned changes may be accompanied with the
unplanned changes in code formatting.

So to avoid this inconvenience, code formatting updated using quite recent
version of Go tools. This is purely cosmetic change, code functionality
is not affected.