UsbDeviceInfo.MfgAndProduct sting replaced with UsbDeviceInfo.MakeAndModel()
MfgAndProduct depends on two other strings, and this is better to
compute it on demand with the explicit method call rather that
save in structure.

When structure contains pre-computed MfgAndProduct, things tends to loose
synchronization when some data this string depend on appears to be
changed.

The need to be able to change Manufacturer and/or ProductName comes
from the fact, that some devices (namely, Xerox B210, but may be some
others too) fail to report these strings reliably and need to be
reset then re-load this information.