Preliminary Support for HWID-Based Quirks
Previously, quirks were selected based on the device model name.
However, some devices (currently, the only known example is the Xerox
B210, but this list may expand in the future) require certain actions to
be performed before the model name can be retrieved from the device.

To address this, we need a way to define quirk rules based on the USB
hardware ID (HWID).

This commit introduces a new syntax for quirks:

  [VVVV:PPPP]
    init-reset = true

or

  [VVVV:*]
    init-reset = true

Here, VVVV and PPPP are exactly 4 hexadecimal digits (0-9, a-f, or A-F),
where:

  * VVVV = USB Vendor ID
  * PPPP = USB Product ID

(Both IDs can be obtained via lsusb.)

Quirks defined in this format will be referred to as HWID quirks.

Additionally, two new quirks are introduced:

  * mfg = "Manufacturer name"
  * model = "Product name"

These quirks are exclusive to the HWID format and allow overriding the
manufacturer and model names in cases where the device cannot reliably
provide them.

Note: This commit only implements the new syntax. The actual
functionality will be added in subsequent commits.